requirements.txt
pandas
torch
transformers
numpy
scikit-learn
matplotlib
seaborn
google-generativeai
openai
dotenv
flask
dropbox

